Updated the Project GUI
The GUI now opens up an available com port for communication with the board.  It also has some error checking for input in the text box.  Also, the speed and pitch boxes have been removed in lieu of a currently transmitted character block that shows the character being transmitted.  When the user hits the send to board button the GUI will lock down the UI and send the string one character at a time to the board with a predetermined delay.