Small Changes
Recoded the clock_en into vhdl and added some test bench files to the
repo.  Biggest change is the timing map that was added to the project
gui.  The timing map contains the characters for morse code along with
the timings for the dots and dashes.  It is put into a map object and
the dot, dash, and gap times can be changed if we need to.