Updated the some testing
I updated the code for the cycle count and shift register test benches.
They both work now and I have images saved for their runs.  These images
can be used in the report.  I need to create a test bench run for the
rec_control_logic and that should be good.